+++ src/function/variables.ts
@@ -155,7 +155,7 @@ export function replaceVariables(variables: Record<string, any>, option: Variabl
     }
     case 'character': {
       const store = useCharacterSettingsStore();
-      if (store.name === undefined) {
+      if (store.avatar === undefined) {
         throw new Error('当前没有打开角色卡，保存角色卡变量失败');
       }
       toRef(store.settings, 'variables').value = variables;